+++ src/model/document.ts
@@ -16,6 +16,7 @@ import {AnalysisContext} from '../core/analysis-context';
 import {ParsedDocument} from '../parser/document';
 import {Behavior} from '../polymer/behavior';
 import {DomModule} from '../polymer/dom-module-scanner';
+import {Namespace} from '../polymer/namespace';
 import {PolymerElement} from '../polymer/polymer-element';
 import {PolymerElementMixin} from '../polymer/polymer-element-mixin';
 import {Warning} from '../warning/warning';
@@ -85,6 +86,7 @@ export interface FeatureKinds {
   'polymer-element': PolymerElement;
   'polymer-element-mixin': PolymerElementMixin;
   'behavior': Behavior;
+  'namespace': Namespace;
   'dom-module': DomModule;
   'element-reference': ElementReference;
   'import': Import;