+++ src/test/static/namespaces/namespace-dynamic-named.js
@@ -1,0 +1,13 @@
+/**
+ * @namespace DynamicNamespace.ArrayNotation
+ */
+DynamicNamespace['ArrayNotation'] = {
+  foo: 'bar'
+};
+
+/**
+ * @namespace DynamicNamespace.Aliased
+ */
+aliasToNamespace = {
+  foo: 'bar'
+};