+++ src/test/static/namespaces/namespace-dynamic-unnamed.js
@@ -1,0 +1,6 @@
+/**
+ * @namespace
+ */
+DynamicNamespace['ArrayNotation'] = {
+  foo: 'bar'
+};